+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Restore the correct parameter for the REST client space URL
 ### Changed
 - SDK now uses `signalwire.receive` to set up contexts
+- The `call.*` actions are now `calling.*`
 
 ## [2.1.2] - 2019-08-01
 ### Fixed